requirements.txt
requests
beautifulsoup4
pandas
python-dotenv
pymongo
dnspython
certifi
openai
tqdm
selenium
webdriver-manager
